+++ src/routes/NotFound/NotFound.jsx
@@ -1,8 +1,6 @@
+import { PageTemplate, Typography, Transition, Button } from "../../components";
+
 import styles from "./NotFound.module.css";
-import PageTemplate from "../../components/PageTemplate/PageTemplate";
-import Typography from "../../components/Typography/Typography";
-import Transition from "../../components/Transition/Transition";
-import Button from "../../components/Button/Button";
 
 function NotFound() {
   return (
