+++ src/routes/Projects/Openings/Openings.jsx
@@ -1,17 +1,19 @@
 import { useLenis } from "lenis/react";
-import Transition from "../../../components/Transition/Transition";
+import { Link } from "react-router-dom";
 import useFetch from "../../../hooks/useFetch";
 import { API_DOMAIN } from "../../../utils/constants";
-import PageTemplate from "../../../components/PageTemplate/PageTemplate";
-import Typography from "../../../components/Typography/Typography";
-import BackButton from "../../../components/BackButton/BackButton";
-import Grid from "../../../components/Grid/Grid";
-import Button from "../../../components/Button/Button";
-import { Link } from "react-router-dom";
-import Image from "../../../components/Image/Image";
+import {
+  PageTemplate,
+  Typography,
+  BackButton,
+  Grid,
+  Button,
+  Transition,
+  Image,
+} from "../../../components";
 
-import styles from "./Openings.module.css";
 import cardStyles from "../../../components/PhotoCard/PhotoCard.module.css";
+import styles from "./Openings.module.css";
 
 const ProjectCard = ({ image, topText, bottomText, to, isRecruiting }) => {
   const Comp = to ? Link : "div";
