--- src/routes/Recruitment/Tinkering/TinkeringProject.module.css
@@ -1,190 +1,0 @@
-.content-wrapper {
-  display: flex;
-  flex-direction: column;
-  row-gap: 240px;
-}
-
-.section-wrapper {
-  display: flex;
-  flex-direction: column;
-  row-gap: var(--content-gap-desktop);
-}
-
-.text-section {
-  display: flex;
-  flex-direction: column;
-  row-gap: var(--para-gap-desktop);
-}
-
-.banner {
-  height: calc(100vh - var(--top-padding));
-  min-height: 760px;
-  display: flex;
-  align-items: flex-start;
-  flex-direction: column;
-  row-gap: 1.5rem;
-  padding-bottom: 1.5rem;
-}
-
-.banner-space {
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-end;
-  width: 100%;
-}
-
-.banner-image {
-  min-height: 640px;
-}
-
-.tinkererNight {
-  display: flex;
-  gap: var(--content-gap-mobile);
-  align-items: center;
-}
-
-.tablet {
-  display: none;
-}
-
-.tinkererNight-image {
-  border-radius: 2rem;
-  width: 100%;
-  height: 100%;
-}
-
-.join-text,
-.garage-assigned-text,
-.programme-section {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  row-gap: 40px;
-  align-items: left;
-}
-
-.join-text .text-section button {
-  color: #fff;
-  width: 100%;
-  text-align: center;
-  padding: 1rem;
-  background-color: var(--garage-blue);
-}
-
-.join-text .text-section button:disabled {
-  background-color: #cccccc;
-}
-
-.project-cards-bottom-text,
-.support-cards-bottom-text {
-  text-align: center;
-}
-
-.loading-wrapper {
-  width: 100%;
-  height: 180px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.custom-photo-layout {
-  display: grid;
-  gap: 20px;
-  grid-template-areas:
-    "left1 center right1"
-    "left2 center right2";
-  grid-template-columns: 1fr 2.3fr 1fr;
-}
-
-.custom-photo-layout > :nth-child(1) {
-  grid-area: left1;
-}
-.custom-photo-layout > :nth-child(2) {
-  grid-area: left2;
-}
-.custom-photo-layout > :nth-child(3) {
-  grid-area: center;
-}
-.custom-photo-layout > :nth-child(4) {
-  grid-area: right1;
-}
-.custom-photo-layout > :nth-child(5) {
-  grid-area: right2;
-}
-
-.sharing-cards-bottom-text {
-  text-align: center;
-  font-size: 20px;
-  line-height: 28px;
-  font-weight: 400;
-}
-
-@media (max-width: 1200px) {
-  .scroll-more {
-    display: none;
-  }
-  .section-wrapper {
-    row-gap: var(--content-gap-tablet);
-  }
-  .text-section {
-    row-gap: var(--para-gap-tablet);
-  }
-  .banner {
-    height: auto;
-    min-height: unset;
-  }
-  .content-wrapper {
-    row-gap: 180px;
-  }
-  .tinkererNight {
-    flex-direction: column;
-    gap: 24px;
-    align-items: flex-start;
-  }
-  .tablet {
-    display: unset;
-  }
-  .tablet-hide {
-    display: none;
-  }
-  .custom-photo-layout {
-    grid-template-areas:
-      "left1 left2"
-      "center center"
-      "right1 right2";
-    grid-template-columns: 1fr 1fr;
-  }
-  .sharing-cards-bottom-text {
-    font-size: 18px;
-    line-height: 28px;
-  }
-}
-
-@media (max-width: 640px) {
-  .section-wrapper {
-    row-gap: var(--content-gap-mobile);
-  }
-  .text-section {
-    row-gap: var(--para-gap-mobile);
-  }
-  .content-wrapper {
-    row-gap: 90px;
-  }
-  .banner-image {
-    min-height: 320px;
-  }
-  .custom-photo-layout {
-    grid-template-areas:
-      "left1"
-      "left2"
-      "center"
-      "right1"
-      "right2";
-    grid-template-columns: 1fr;
-  }
-  .sharing-cards-bottom-text {
-    font-size: 16px;
-    line-height: 24px;
-  }
-}
